+++ test/test_distrib_pkg.py
@@ -11,6 +11,7 @@ def test_get_no_digests(fragment):
         url="https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl" + fragment,
         requires_python=None,
         has_sig=False,
+        yanked=None,
     )
     assert pkg.get_digests() == {}
 
@@ -23,6 +24,7 @@ def test_get_digests():
         url="https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl#sha256=da69d28dcd527c0e372b3fa7b92fc333b327f8470175f035abc4e351b539189f",
         requires_python=None,
         has_sig=False,
+        yanked=None,
     )
     assert pkg.get_digests() == {
         "sha256": "da69d28dcd527c0e372b3fa7b92fc333b327f8470175f035abc4e351b539189f"
@@ -37,6 +39,7 @@ def test_get_no_sig_url():
         url="https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl#sha256=da69d28dcd527c0e372b3fa7b92fc333b327f8470175f035abc4e351b539189f",
         requires_python=None,
         has_sig=False,
+        yanked=None,
     )
     assert pkg.sig_url is None
 
@@ -49,5 +52,6 @@ def test_get_sig_url():
         url="https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl#sha256=da69d28dcd527c0e372b3fa7b92fc333b327f8470175f035abc4e351b539189f",
         requires_python=None,
         has_sig=True,
+        yanked=None,
     )
     assert pkg.sig_url == "https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl.asc"